+++ lib/ModelMixins/GeojsonMixin.ts
@@ -10,7 +10,8 @@ import {
   MultiPolygon,
   Point,
   Polygon,
-  Properties
+  Properties,
+  MultiPoint
 } from "@turf/helpers";
 import i18next from "i18next";
 import {
@@ -180,9 +181,34 @@ class GeoJsonStratum extends LoadableStratum(GeoJsonTraits) {
   get showDisableStyleOption() {
     return true;
   }
+
+  @computed get forceCesiumPrimitives() {
+    // Disable TableStyling for the following:
+    // If MultiPoint features exist
+    // If more than 50% of features have simple style properties - disable table styling
+    if (
+      this._item.featureCounts.multiPoint > 0 ||
+      this._item.featureCounts.simpleStyle / this._item.featureCounts.total >=
+        0.5
+    ) {
+      return true;
+    }
+  }
 }
 
 StratumOrder.addLoadStratum(GeoJsonStratum.stratumName);
+
+interface FeatureCounts {
+  point: number;
+  multiPoint: number;
+  /** Line includes MultiLine features */
+  line: number;
+  /** Polygon includes MultiPolygon features */
+  polygon: number;
+  /** Count of features with simplestyle-spec properties (eg "fill-color") */
+  simpleStyle: number;
+  total: number;
+}
 
 function GeoJsonMixin<T extends Constructor<Model<GeoJsonTraits>>>(Base: T) {
   abstract class GeoJsonMixin extends TableMixin(
@@ -205,11 +231,14 @@ function GeoJsonMixin<T extends Constructor<Model<GeoJsonTraits>>>(Base: T) {
     @observable.ref _readyData?: FeatureCollectionWithCrs;
 
     /** Number of features in _readyData FeatureCollection */
-    @observable featureCounts: {
-      point: number;
-      line: number;
-      polygon: number;
-    } = { point: 0, line: 0, polygon: 0 };
+    @observable featureCounts: FeatureCounts = {
+      point: 0,
+      multiPoint: 0,
+      line: 0,
+      polygon: 0,
+      simpleStyle: 0,
+      total: 0
+    };
 
     constructor(...args: any[]) {
       super(...args);
@@ -408,10 +437,10 @@ function GeoJsonMixin<T extends Constructor<Model<GeoJsonTraits>>>(Base: T) {
      * - Cesium primitives if:
      *    - `GeoJsonTraits.forceCesiumPrimitives = true`
      *    - Using `timeProperty` or `heightProperty` or `perPropertyStyles` or simple-style `marker-symbol`
-     *    - More than 50% of GeoJSON features have simply-style properties
+     *    - More than 50% of GeoJSON features have simply-style properties (eg "fill-color")
+     *    - MultiPoint features are in GeoJSON (not supported by Table styling)
      */
     protected async forceLoadMapItems(): Promise<void> {
-      let useTableStylingAndProtomaps = this.useTableStylingAndProtomaps;
       const czmlTemplate = this.czmlTemplate;
       const filterByProperties = this.filterByProperties;
 
@@ -428,12 +457,14 @@ function GeoJsonMixin<T extends Constructor<Model<GeoJsonTraits>>>(Base: T) {
           this.terria.configParameters.proj4ServiceBaseUrl
         );
 
-        // Add feature index to FEATURE_ID_PROP ("_id_") feature property
-        // This is used to refer to each feature in TableMixin (as row ID)
-
-        // Also check for how many features have simply-style properties
-        let numFeaturesWithSimpleStyle = 0;
-        const featureCounts = { point: 0, line: 0, polygon: 0 };
+        const featureCounts: FeatureCounts = {
+          point: 0,
+          multiPoint: 0,
+          line: 0,
+          polygon: 0,
+          simpleStyle: 0,
+          total: 0
+        };
 
         // We will re-add features depending if filterByProperties - or geometry is invalid
         const features = geoJsonWgs84.features;
@@ -462,14 +493,16 @@ function GeoJsonMixin<T extends Constructor<Model<GeoJsonTraits>>>(Base: T) {
 
           geoJsonWgs84.features.push(feature);
 
+          // Add feature index to FEATURE_ID_PROP ("_id_") feature property
+          // This is used to refer to each feature in TableMixin (as row ID)
           const properties = feature.properties!;
           properties[FEATURE_ID_PROP] = i;
 
-          if (
-            feature.geometry.type === "Point" ||
-            feature.geometry.type === "MultiPoint"
-          ) {
+          // Count features types
+          if (feature.geometry.type === "Point") {
             featureCounts.point++;
+          } else if (feature.geometry.type === "MultiPoint") {
+            featureCounts.multiPoint++;
           } else if (
             feature.geometry.type === "LineString" ||
             feature.geometry.type === "MultiLineString"
@@ -482,28 +515,16 @@ function GeoJsonMixin<T extends Constructor<Model<GeoJsonTraits>>>(Base: T) {
             featureCounts.polygon++;
           }
 
-          if (
-            useTableStylingAndProtomaps &&
-            SIMPLE_STYLE_KEYS.find(key => properties[key])
-          ) {
-            numFeaturesWithSimpleStyle++;
+          // Does feature include simplestyle-spec properties (eg "fill-colour)")
+          if (SIMPLE_STYLE_KEYS.find(key => properties[key])) {
+            featureCounts.simpleStyle++;
           }
+
+          featureCounts.total++;
         }
 
-        runInAction(() => (this.featureCounts = featureCounts));
-
-        // If more than 50% of features have simple style properties - disable table styling
-        if (numFeaturesWithSimpleStyle / geoJsonWgs84.features.length >= 0.5) {
-          runInAction(() => {
-            this.setTrait(
-              CommonStrata.underride,
-              "forceCesiumPrimitives",
-              true
-            );
-            useTableStylingAndProtomaps = this.useTableStylingAndProtomaps;
-          });
-        }
         runInAction(() => {
+          this.featureCounts = featureCounts;
           this._readyData = geoJsonWgs84;
         });
 
@@ -513,7 +534,7 @@ function GeoJsonMixin<T extends Constructor<Model<GeoJsonTraits>>>(Base: T) {
             this._dataSource = dataSource;
             this._imageryProvider = undefined;
           });
-        } else if (useTableStylingAndProtomaps) {
+        } else if (runInAction(() => this.useTableStylingAndProtomaps)) {
           runInAction(() => {
             this._imageryProvider = this.createProtomapsImageryProvider(
               geoJsonWgs84
@@ -649,6 +670,8 @@ function GeoJsonMixin<T extends Constructor<Model<GeoJsonTraits>>>(Base: T) {
       const rows = this.activeTableStyle.colorColumn?.valuesForType;
       const colorMap = this.activeTableStyle.colorMap;
       const outlineStyleMap = this.activeTableStyle.outlineStyleMap.styleMap;
+      const useOutlineColorForLineFeatures = this
+        .useOutlineColorForLineFeatures;
 
       // Style function
       const getColorValue = (z: number, f?: ProtomapsFeature) => {
@@ -707,8 +730,7 @@ function GeoJsonMixin<T extends Constructor<Model<GeoJsonTraits>>>(Base: T) {
             dataLayer: GEOJSON_SOURCE_LAYER_NAME,
             symbolizer: new PolygonSymbolizer({
               fill: getColorValue,
-              stroke: getOutlineColorValue,
-              width: getOutlineWidthValue
+              stroke: getOutlineColorValue
             }),
             minzoom: 0,
             maxzoom: Infinity,
@@ -719,11 +741,17 @@ function GeoJsonMixin<T extends Constructor<Model<GeoJsonTraits>>>(Base: T) {
               );
             }
           },
+
           // Line features
+          // Note - line color will use TableColorStyleTraits by default.
+          // If useOutlineColorForLineFeatures is true, then line color will use TableOutlineStyle traits
+
           {
             dataLayer: GEOJSON_SOURCE_LAYER_NAME,
             symbolizer: new LineSymbolizer({
-              color: getColorValue,
+              color: useOutlineColorForLineFeatures
+                ? getOutlineColorValue
+                : getColorValue,
               width: getOutlineWidthValue
             }),
             minzoom: 0,
@@ -1228,6 +1256,14 @@ export function isFeature(json: any): json is Feature {
 export function isPoint(json: any): json is Feature<Point> {
   return (
     json.type === "Feature" && json.geometry && json.geometry.type === "Point"
+  );
+}
+
+export function isMultiPoint(json: any): json is Feature<MultiPoint> {
+  return (
+    json.type === "Feature" &&
+    json.geometry &&
+    json.geometry.type === "MultiPoint"
   );
 }