+++ CHANGES.md
@@ -3,6 +3,9 @@ Change Log
 
 #### next release (8.2.8)
 
+* Table styling is disabled if `MultiPoint` are in GeoJSON
+* Add `GeoJsonTraits.useOutlineColorForLineFeatures` - If enabled, `TableOutlineStyleTraits` will be used to color Line Features, otherwise `TableColorStyleTraits` will be used.
+* Fix feature highliting for `Line`, `MultiLine` and `MultiPoint`
 * Await Internationalisation initialisation in `Terria.start`
 * `UserDrawing.messageHeader` can now also be `() => string`
 * Protomaps Polygon features now only use `PolygonSymbolizer` (instead of `PolygonSymbolizer` and `LineSymbolizer`)
@@ -37,7 +40,6 @@ Change Log
 * Added a required parameter i18n to `applyTranslationIfExists` to avoid having stale translated strings when the language changes.
 * Fix `StoryBuilder` remove all text color
 * Fix `FeatureInfoPanel` `Loader` color
-* [The next improvement]
 
 #### 8.2.6 - 2022-06-17
 
@@ -51,7 +53,6 @@ Change Log
 * Made `order` optional for `ICompositeBarItem`.
 * Fix `includes` path for `url-loader` rule so that it doesn't incorrectly match package names with `terriajs` as prefix.
 * Add help button for bookmarking sharelinks to SharePanel (if that help item exists in config)
-* [The next improvement]
 
 #### 8.2.5 - 2022-06-07